+++ packages/zowe-explorer/__tests__/__unit__/uss/ZoweUSSNode.unit.test.ts
@@ -12,7 +12,7 @@
 import * as vscode from "vscode";
 import * as zosfiles from "@zowe/zos-files-for-zowe-sdk";
 import * as zosmf from "@zowe/zosmf-for-zowe-sdk";
-import { Gui, imperative, Validation, ZoweScheme } from "@zowe/zowe-explorer-api";
+import { Gui, imperative, UssDirectory, UssFile, Validation, ZoweScheme } from "@zowe/zowe-explorer-api";
 import { ZoweExplorerApiRegister } from "../../../src/ZoweExplorerApiRegister";
 import { Profiles } from "../../../src/Profiles";
 import { ZoweUSSNode } from "../../../src/uss/ZoweUSSNode";
@@ -436,6 +436,21 @@ describe("ZoweUSSNode Unit Tests - Function node.rename()", () => {
         newMocks.ussDir.contextValue = globals.USS_DIR_CONTEXT;
         return newMocks;
     }
+
+    it("Tests that when rename fails, an error message is thrown", async () => {
+        const globalMocks = await createGlobalMocks();
+        const blockMocks = await createBlockMocks(globalMocks);
+
+        const newFullPath = "/u/user/newName";
+        const errMessageMock = jest.spyOn(Gui, "errorMessage").mockImplementation();
+        const renameMock = jest.spyOn(UssFSProvider.instance, "rename").mockRejectedValueOnce(new Error("Rename error: file is busy"));
+        await blockMocks.ussDir.rename(newFullPath);
+
+        expect(errMessageMock).toHaveBeenCalledWith("Rename error: file is busy");
+        errMessageMock.mockRestore();
+        renameMock.mockRestore();
+    });
+
     it("Tests that rename updates and refreshes the UI components of the node", async () => {
         const globalMocks = await createGlobalMocks();
         const blockMocks = await createBlockMocks(globalMocks);
@@ -445,7 +460,6 @@ describe("ZoweUSSNode Unit Tests - Function node.rename()", () => {
 
         // Expect renamed node's labels to be updated with newName
         expect(blockMocks.ussDir.fullPath).toEqual(newFullPath);
-        expect(blockMocks.ussDir.shortLabel).toEqual("newName");
         expect(blockMocks.ussDir.label).toEqual("newName");
         expect(blockMocks.ussDir.tooltip).toEqual(newFullPath);
 
@@ -477,7 +491,6 @@ describe("ZoweUSSNode Unit Tests - Function node.rename()", () => {
             parentPath: "/u/user/ussDir/ussSubDir",
         });
         ussSubDirChild.contextValue = globals.USS_TEXT_FILE_CONTEXT;
-        ussSubDirChild.shortLabel = "ussSubDirChild/ussChildFile";
         ussSubDir.children.push(ussSubDirChild);
 
         const newFullPath = "/u/user/newName";
@@ -534,7 +547,7 @@ describe("ZoweUSSNode Unit Tests - Function node.reopen()", () => {
     });
 });
 
-describe("ZoweUSSNode Unit Tests - Function node.setEncoding()", () => {
+describe("ZoweUSSNode Unit Tests - node.setEncoding() and encoding behaviors", () => {
     const setEncodingForFileMock = jest.spyOn(UssFSProvider.instance, "setEncodingForFile").mockImplementation();
     const getEncodingMock = jest.spyOn(UssFSProvider.instance as any, "getEncodingForFile");
 
@@ -551,7 +564,6 @@ describe("ZoweUSSNode Unit Tests - Function node.setEncoding()", () => {
         getEncodingMock.mockReturnValue(binaryEncoding);
         const node = new ZoweUSSNode({ label: "encodingTest", collapsibleState: vscode.TreeItemCollapsibleState.None });
         node.setEncoding(binaryEncoding);
-        expect(node.binary).toEqual(true);
         expect(setEncodingForFileMock).toHaveBeenCalledWith(node.resourceUri, binaryEncoding);
         expect(node.tooltip).toContain("Encoding: Binary");
         expect(node.contextValue).toEqual(globals.USS_BINARY_FILE_CONTEXT);
@@ -563,7 +575,6 @@ describe("ZoweUSSNode Unit Tests - Function node.setEncoding()", () => {
         getEncodingMock.mockReturnValue(textEncoding);
         const node = new ZoweUSSNode({ label: "encodingTest", collapsibleState: vscode.TreeItemCollapsibleState.None });
         node.setEncoding(textEncoding);
-        expect(node.binary).toEqual(false);
         expect(setEncodingForFileMock).toHaveBeenCalledWith(node.resourceUri, textEncoding);
         expect(node.tooltip).not.toContain("Encoding:");
         expect(node.contextValue).toEqual(globals.USS_TEXT_FILE_CONTEXT);
@@ -574,7 +585,6 @@ describe("ZoweUSSNode Unit Tests - Function node.setEncoding()", () => {
         getEncodingMock.mockReturnValue(otherEncoding);
         const node = new ZoweUSSNode({ label: "encodingTest", collapsibleState: vscode.TreeItemCollapsibleState.None });
         node.setEncoding(otherEncoding);
-        expect(node.binary).toEqual(false);
         expect(setEncodingForFileMock).toHaveBeenCalledWith(node.resourceUri, otherEncoding);
         expect(node.tooltip).toContain("Encoding: IBM-1047");
     });
@@ -589,7 +599,6 @@ describe("ZoweUSSNode Unit Tests - Function node.setEncoding()", () => {
         getEncodingMock.mockReturnValue(textEncoding);
         const node = new ZoweUSSNode({ label: "encodingTest", collapsibleState: vscode.TreeItemCollapsibleState.None, parentNode });
         node.setEncoding(textEncoding);
-        expect(node.binary).toEqual(false);
         expect(setEncodingForFileMock).toHaveBeenCalledWith(node.resourceUri, textEncoding);
         expect(node.contextValue).toEqual(globals.USS_TEXT_FILE_CONTEXT + globals.FAV_SUFFIX);
     });
@@ -598,7 +607,6 @@ describe("ZoweUSSNode Unit Tests - Function node.setEncoding()", () => {
         const node = new ZoweUSSNode({ label: "encodingTest", collapsibleState: vscode.TreeItemCollapsibleState.None });
         getEncodingMock.mockReturnValue(undefined);
         node.setEncoding(undefined as any);
-        expect(node.binary).toEqual(false);
         expect(setEncodingForFileMock).toHaveBeenCalledWith(node.resourceUri, undefined);
     });
 
@@ -719,6 +727,8 @@ describe("ZoweUSSNode Unit Tests - Function node.getChildren()", () => {
         blockMocks.rootNode.contextValue = globals.USS_DIR_CONTEXT;
         blockMocks.rootNode.dirty = true;
 
+        const setAttrsMock = jest.spyOn(ZoweUSSNode.prototype, "setAttributes").mockImplementation();
+
         // Creating structure of files and directories
         const sampleChildren: ZoweUSSNode[] = [
             new ZoweUSSNode({
@@ -749,6 +759,7 @@ describe("ZoweUSSNode Unit Tests - Function node.getChildren()", () => {
         expect(rootChildren.length).toBe(2);
         expect(rootChildren[0].label).toBe("aDir");
         expect(rootChildren[1].label).toBe("myFile.txt");
+        setAttrsMock.mockRestore();
     });
 
     it("Tests that node.getChildren() returns no children if none exist", async () => {
@@ -776,6 +787,8 @@ describe("ZoweUSSNode Unit Tests - Function node.getChildren()", () => {
         const blockMocks = await createBlockMocks(globalMocks);
         const oldPath = "/u/oldUser";
         const newPath = "/u/newUser";
+
+        const setAttrsMock = jest.spyOn(ZoweUSSNode.prototype, "setAttributes").mockImplementation();
 
         const parentNode = new ZoweUSSNode({
             label: "newUser",
@@ -814,6 +827,8 @@ describe("ZoweUSSNode Unit Tests - Function node.getChildren()", () => {
         expect(newChildren[1].fullPath).not.toContain(oldPath);
         expect(newChildren[0].fullPath).toContain(newPath);
         expect(newChildren[1].fullPath).toContain(newPath);
+        expect(setAttrsMock).toHaveBeenCalled();
+        setAttrsMock.mockRestore();
     });
 
     it("Tests that error is thrown when node label is blank", async () => {
@@ -1364,5 +1379,77 @@ describe("ZoweUSSNode Unit Tests - Function node.pasteUssTree()", () => {
         jest.spyOn(blockMocks.mockUssApi, "uploadDirectory").mockRejectedValueOnce(blockMocks.fileResponse);
 
         await blockMocks.testNode.pasteUssTree();
+    });
+});
+
+describe("ZoweUSSNode Unit Tests - Function node.setEtag", () => {
+    it("sets the e-tag for a file", () => {
+        const fileEntry = new UssFile("testFile");
+        const statMock = jest.spyOn(UssFSProvider.instance, "stat").mockReturnValueOnce(fileEntry);
+
+        const node = new ZoweUSSNode({ label: "testFile", collapsibleState: vscode.TreeItemCollapsibleState.None });
+        node.setEtag("123ETAG");
+        expect(statMock).toHaveBeenCalled();
+        expect(fileEntry.etag).toBe("123ETAG");
+        statMock.mockRestore();
+    });
+
+    it("returns early when trying to set the e-tag for a directory", () => {
+        const dirEntry = new UssDirectory("testDir");
+        const statMock = jest.spyOn(UssFSProvider.instance, "stat").mockReturnValueOnce(dirEntry);
+
+        const node = new ZoweUSSNode({ label: "testDir", collapsibleState: vscode.TreeItemCollapsibleState.Collapsed });
+        node.setEtag("123ETAG");
+        expect(statMock).toHaveBeenCalled();
+        expect(dirEntry).not.toHaveProperty("etag");
+        statMock.mockRestore();
+    });
+});
+
+describe("ZoweUSSNode Unit Tests - Function node.getAttributes", () => {
+    const attrs = { owner: "aUser", uid: 0, gid: 1000, group: "USERS", perms: "rwxrwxrwx" };
+    it("gets the attributes for a file", () => {
+        const fileEntry = new UssFile("testFile");
+        fileEntry.attributes = attrs;
+        const statMock = jest.spyOn(UssFSProvider.instance, "stat").mockReturnValueOnce(fileEntry);
+
+        const node = new ZoweUSSNode({ label: "testFile", collapsibleState: vscode.TreeItemCollapsibleState.None });
+        expect(node.getAttributes()).toStrictEqual(attrs);
+        statMock.mockRestore();
+    });
+});
+
+describe("ZoweUSSNode Unit Tests - Function node.setAttributes", () => {
+    const attrs = { owner: "aUser", uid: 0, gid: 1000, group: "USERS", perms: "rwxrwxrwx" };
+    it("sets the attributes for a file", () => {
+        const fileEntry = new UssFile("testFile");
+        fileEntry.attributes = { ...attrs };
+        const statMock = jest.spyOn(UssFSProvider.instance, "stat").mockReturnValueOnce(fileEntry);
+
+        const node = new ZoweUSSNode({ label: "testFile", collapsibleState: vscode.TreeItemCollapsibleState.None });
+        node.setAttributes({ perms: "r-xr-xr-x" });
+        expect(statMock).toHaveBeenCalled();
+        expect(fileEntry.attributes).toStrictEqual({ ...attrs, perms: "r-xr-xr-x" });
+        statMock.mockRestore();
+    });
+
+    it("sets the attributes for a directory", () => {
+        const dirEntry = new UssDirectory("testFolder");
+        const statMock = jest.spyOn(UssFSProvider.instance, "stat").mockClear().mockReturnValueOnce(dirEntry);
+        dirEntry.attributes = { ...attrs };
+
+        const node = new ZoweUSSNode({ label: "testFolder", collapsibleState: vscode.TreeItemCollapsibleState.Collapsed });
+        node.setAttributes({ perms: "r-xr-xr-x" });
+        expect(statMock).toHaveBeenCalled();
+        expect(dirEntry.attributes).toStrictEqual({ ...attrs, perms: "r-xr-xr-x" });
+        statMock.mockRestore();
+    });
+});
+
+describe("ZoweUSSNode Unit Tests - Function node.getBaseName", () => {
+    it("returns the base name for a USS node based on its URI", async () => {
+        const node = new ZoweUSSNode({ label: "testFile", collapsibleState: vscode.TreeItemCollapsibleState.None });
+        node.resourceUri = vscode.Uri.from({ scheme: ZoweScheme.USS, path: "/someProfile/a/b/c/testFile" });
+        expect(node.getBaseName()).toBe("testFile");
     });
 });